fix(web): app layout showing for logged out users
To fix this, this actually does a whole lot refactor in the app routes.
Now we have a `_authenticated/` (pathless route, i.e it does not appear
in the address bar) folder that groups all the app routes that requires
authentication; We also have a `_public/` (also pathless) folder to
group the public routes, `/sign-in` and `/sign-up`, and perhaps more in
the future.