fix(web): change trpc wrapper url
By using `import.meta.env.VITE_API_URL` we would be fine for localhost
but it wouldn't work when using different urls for referencing the base
app (e.g: accessing it in the local wi-fi network for instance).

Using `window.location.origin` we solve that given the assumption we're
always pointing to an API hosted in the same domain + we don't need a
differentiation between test vs dev env.